ads_ref_extract/tex.py: try to deal with unusual linewrapping in bibitems
Some bibitems were being munged incorrectly because they were split over
lines in ways unexpected by the original implementation. In particular,
it looks like some people have some kind of formatter that likes to
split items using trailing comment characters (%), taking advantage of
TeX's whitespace munching after comments.

We can increase the success rate of the munging by eliminating these
comments. But, of course, it's never that simple. First, there are times
when a percent sign isn't a comment: when it's part of `\%`, or in a URL
if the relevant code futzes with its catcode. We detect the latter case
by checking of the % is followed by two hex-like characters, which feels
super sketchy but seems to actually work.

It all feels a bit fragile, but any line-based TeX parser is going to be
like that. Empirically, this tweak seems to lead to good results in my
test cases.